refactor(e2e): move assertDemoClientSession to support/utils
The helper was defined privately inside auth.steps.ts but needs to be
reusable from consent.steps.ts too, now that consent-screen scenarios
distinguish trusted and untrusted demo clients and want to assert
which demo the user landed on after a valid session.

Lift it into e2e/support/utils.ts and add an optional expectedDemoUrl
parameter that asserts the post-redirect URL's origin matches the given
demo client. Omitting the parameter preserves the existing behaviour
(any demo origin accepted), so the existing "redirected back to the
demo client with a valid session" step keeps working unchanged.
