+++ e2e/support/utils.ts
@@ -2,7 +2,7 @@
  * Shared utility helpers for step definitions.
  */
 
-import type { Browser } from '@playwright/test'
+import { expect, type Browser } from '@playwright/test'
 import type { EpdsWorld } from './world.js'
 import { testEnv } from './env.js'
 
@@ -34,6 +34,49 @@ export async function resetBrowserContext(
   world.page = await world.context.newPage()
   world.page.setDefaultNavigationTimeout(30_000)
   world.page.setDefaultTimeout(15_000)
+}
+
+/**
+ * Asserts that the browser has landed on a demo client's /welcome page
+ * with a valid session. Checks: /welcome URL pattern, session_id cookie
+ * set, body text indicating signed-in state, visible DID and handle.
+ *
+ * Pass `expectedDemoUrl` to additionally assert the welcome URL's origin
+ * matches that demo client's origin — this is how consent-screen scenarios
+ * distinguish "redirected to the trusted demo" from "redirected to the
+ * untrusted demo" when both demo services are live in the test env.
+ * Omit it to accept any demo origin (back-compat for scenarios that only
+ * exercise a single demo client).
+ */
+export async function assertDemoClientSession(
+  world: EpdsWorld,
+  expectedDemoUrl?: string,
+): Promise<void> {
+  const page = getPage(world)
+
+  await page.waitForURL('**/welcome', { timeout: 30_000 })
+
+  if (expectedDemoUrl) {
+    const expectedOrigin = new URL(expectedDemoUrl).origin
+    const actualOrigin = new URL(page.url()).origin
+    if (actualOrigin !== expectedOrigin) {
+      throw new Error(
+        `Expected redirect to ${expectedOrigin}, got ${actualOrigin}`,
+      )
+    }
+  }
+
+  const cookies = await page.context().cookies()
+  const sessionCookie = cookies.find((cookie) => cookie.name === 'session_id')
+  if (!sessionCookie?.value) {
+    throw new Error('Demo client session cookie was not set after redirect')
+  }
+
+  const body = page.locator('body')
+  await expect(body).toContainText('You are signed in.')
+  await expect(body).toContainText('Sign out')
+  await expect(body).toContainText(/@[\w.-]+/)
+  await expect(body).toContainText(/did:[a-z0-9:]+/i)
 }
 
 /**
